apps/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nstil"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "supabase>=2.13",
    "redis[hiredis]>=5.2",
    "arq>=0.26",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.20",
    "httpx>=0.28",
    "structlog>=25.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
    "mypy>=1.14",
    "httpx>=0.28",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nstil"]

[tool.ruff]
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "N"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
mypy_path = "src"
packages = ["nstil"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = [".", "src"]
